Fix sbt 0.13.13 warnings, upgrade deps
