+++ .jvmopts
@@ -1,7 +1,6 @@
 -Dfile.encoding=UTF8
 -Xms512m
 -Xmx2g
--XX:MaxPermSize=1g
 -Xss4m
 -XX:ReservedCodeCacheSize=128m
 -Djava.awt.headless=true
