+++ publishing.sbt
@@ -1,6 +1,4 @@
-publishTo in ThisBuild <<= (version) { version: String =>
-  if (version.trim.endsWith("SNAPSHOT"))
-    Some(Opts.resolver.sonatypeSnapshots)
-  else
-    Some(Opts.resolver.sonatypeStaging)
+publishTo in ThisBuild := {
+  if (version.value.trim.endsWith("SNAPSHOT")) Some(Opts.resolver.sonatypeSnapshots)
+  else Some(Opts.resolver.sonatypeStaging)
 }
